Decorative patten test One

+++ src/decorating/object/one/Beverage.java
@@ -8,7 +8,7 @@ public abstract class Beverage {
     private String description;
 
 
-    abstract int getCost();
+    abstract float getCost();
 
 
     public String getDescription() {
