+++ src/decorating/object/one/DarkRoast.java
@@ -7,7 +7,7 @@ public class DarkRoast extends Beverage {
 
 
     @Override
-    int getCost() {
+    float getCost() {
         return 0;
     }
 }
